Create StringCFGVisualizer, minimize the duplication between DOTCFGVisualizer and StringVisualizer.
